requirements.txt
tensorflow-probability
tensorflow-datasets
tensorflow-addons

numpy
scipy
wandb
hydra-core
hydra_colorlog
hydra-optuna-sweeper
omegaconf
cloud-tpu-client

black
rich